% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_extra.R
\name{mean_contrasts}
\alias{mean_contrasts}
\title{Display contrasts}
\usage{
mean_contrasts(M)
}
\arguments{
\item{M}{Any \eqn{n \times (n-1)}{n x (n-1)} coding matrix.}
}
\value{
The full contrast matrix, \code{solve(cbind(1, M))}, suitably
  annotated and presented in vulgar fractional form, for clarity.
}
\description{
A function to display the averaging vector and class mean contrasts
implied by a given coding matrix
}
\examples{
mean_contrasts(code_helmert_forward(5))
mean_contrasts(code_diff_forward(letters[1:7]))
}

